requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.7.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==21.4b2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
coloredlogs==15.0.1
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.35
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
dnspython==2.3.0
effdet==0.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
Faker==18.9.0
fastapi==0.95.1
favicon==0.7.0
ffmpeg==1.4
ffmpy==0.3.1
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
fpdf==1.7.2
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
gradio==3.39.0
gradio_client==0.3.0
greenlet==2.0.2
grpcio==1.54.2
gTTS==2.3.2
h11==0.14.0
htbuilder==0.6.1
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.14.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
iopath==0.1.9
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langchain==0.0.249
langchainplus-sdk==0.0.20
langsmith==0.0.16
layoutparser==0.3.4
linkify-it-py==2.0.2
lit==16.0.3
llvmlite==0.40.0
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.0
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
orjson==3.9.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.1
pymdown-extensions==10.0.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
st-annotated-text==4.0.0
stack-data==0.6.2
starlette==0.26.1
streamlit==1.22.0
streamlit-audiorecorder==0.0.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-chat==0.0.2.2
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
tesseract==0.1.3
tiktoken==0.3.3
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.1
triton==2.0.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzlocal==5.0.1
uc-micro-py==1.0.2
unstructured==0.6.6
unstructured-inference==0.4.4
urllib3==1.26.15
uvicorn==0.22.0
validators==0.20.0
Wand==0.6.11
watchdog==3.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
whisper==1.1.10
wikipedia==1.4.0
wrapt==1.14.1
XlsxWriter==3.1.0
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
